+++ attribution/src/main/java/com/affise/attribution/parameters/providers/InstalledHourProvider.kt
@@ -1,28 +1,24 @@
 package com.affise.attribution.parameters.providers
 
-import android.content.Context
 import com.affise.attribution.parameters.ProviderType
 import com.affise.attribution.parameters.base.LongPropertyProvider
+import com.affise.attribution.usecase.PackageInfoUseCase
 import java.util.Calendar
 import java.util.Date
 
 /**
  * Provider for parameter [ProviderType.INSTALLED_HOUR]
  *
- * @property context to retrieve package manager from
+ * @property useCase to retrieve first install time
  */
 class InstalledHourProvider(
-    private val context: Context
+    private val useCase: PackageInfoUseCase,
 ) : LongPropertyProvider() {
 
     override val order: Float = 8.0f
     override val key: ProviderType = ProviderType.INSTALLED_HOUR
 
-    @Suppress("DEPRECATION")
-    override fun provide(): Long? = context
-        .packageManager
-        .getPackageInfo(context.packageName, 0)
-        ?.firstInstallTime
+    override fun provide(): Long? = useCase.getFirstInstallTime()
         ?.stripTimestampToHours()
 
     private fun Long.stripTimestampToHours() = Calendar.getInstance().apply {