+++ attribution/src/main/java/com/affise/attribution/usecase/FirstAppOpenUseCase.kt
@@ -1,20 +1,19 @@
 package com.affise.attribution.usecase
 
 import android.content.SharedPreferences
-import com.affise.attribution.errors.AffiseError
-import com.affise.attribution.parameters.ProviderType
 import com.affise.attribution.session.CurrentActiveActivityCountProvider
-import com.affise.attribution.utils.*
+import com.affise.attribution.utils.SignType
+import com.affise.attribution.utils.checkSaveString
 import com.affise.attribution.utils.generateUUID
-import com.affise.attribution.utils.timestamp
 import com.affise.attribution.utils.saveBoolean
 import com.affise.attribution.utils.saveLong
-import java.util.*
+import com.affise.attribution.utils.sing
+import com.affise.attribution.utils.timestamp
+import java.util.Date
 
 class FirstAppOpenUseCase(
     private val preferences: SharedPreferences,
     private val activityCountProvider: CurrentActiveActivityCountProvider,
-    private val persistentUseCase: PersistentUseCase
 ) {
 
     private var firstRun: Boolean = false
@@ -52,17 +51,9 @@ class FirstAppOpenUseCase(
 
     private fun checkSaveUUIDs() {
         preferences.apply {
-            //Create affDevId
-            checkSaveString(AFF_DEVICE_ID) {
-                persistentUseCase.getAffDeviceId() ?: generateUUID().toString()
-            }
             //Create affAltDevId
-            checkSaveString(AFF_ALT_DEVICE_ID) {
-                generateUUID().toString()
-            }
-            //Create randomUserId
-            checkSaveString(ProviderType.RANDOM_USER_ID.provider) {
-                generateUUID().toString()
+            checkSaveString(AppUUIDs.AFF_ALT_DEVICE_ID) {
+                generateUUID().toString().sing(SignType.RANDOM)
             }
         }
     }
@@ -98,40 +89,8 @@ class FirstAppOpenUseCase(
         .getLong(FIRST_OPENED_DATE_KEY, 0)
         .let { if (it == 0L) null else Date(it) }
 
-    /**
-     * Get devise id
-     * @return devise id
-     */
-    fun getAffiseDeviseId() = preferences
-        .getString(
-            AFF_DEVICE_ID,
-            AffiseError.ALL_TWO
-        )
-
-    /**
-     * Get alt devise id
-     * @return alt devise id
-     */
-    fun getAffiseAltDeviseId() = preferences
-        .getString(
-            AFF_ALT_DEVICE_ID,
-            AffiseError.ALL_TWO
-        )
-
-    /**
-     * Get random user id
-     * @return random user id
-     */
-    fun getRandomUserId() = preferences
-        .getString(
-            ProviderType.RANDOM_USER_ID.provider,
-            AffiseError.ALL_TWO
-        )
-
     companion object {
         private const val FIRST_OPENED = "FIRST_OPENED"
         private const val FIRST_OPENED_DATE_KEY = "FIRST_OPENED_DATE_KEY"
-        private const val AFF_DEVICE_ID = "AFF_DEVICE_ID"
-        private const val AFF_ALT_DEVICE_ID = "AFF_ALT_DEVICE_ID"
     }
 }