+++ attribution/src/main/java/com/affise/attribution/usecase/PackageInfoUseCase.kt
@@ -1,0 +1,15 @@
+package com.affise.attribution.usecase
+
+interface PackageInfoUseCase {
+
+    fun getFirstInstallTime(): Long?
+
+    fun getAppVersion(): String?
+
+    fun getAppVersionRaw(): String?
+
+    /**
+     * Get initiating app package name
+     */
+    fun getInitiatingPackageName(): String?
+}